% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2_methods.R
\name{Dicker_2013}
\alias{Dicker_2013}
\title{Dicker 2013
the function is based on the Lee H Dicker's 2013 paper: Variance estimation i high-dimensional linear models}
\usage{
Dicker_2013(x, y, target = c("beta2", "h2")[1])
}
\arguments{
\item{x}{A matrix for covariates}

\item{y}{A continuous vector}

\item{target}{A character for estimation target}
}
\description{
Dicker 2013
the function is based on the Lee H Dicker's 2013 paper: Variance estimation i high-dimensional linear models
}
