% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2_methods.R
\name{EigenPrism}
\alias{EigenPrism}
\title{EigenPrism 
Author: Lucas Janson (statweb.stanford.edu/~ljanson)}
\usage{
EigenPrism(
  y,
  x,
  invsqrtSig = NULL,
  alpha = 0.05,
  target = c("beta2", "h2")[1],
  zero.ind = c(),
  diagnostics = F
)
}
\arguments{
\item{y}{A continuous vector}

\item{target}{A character for estimation target}

\item{X}{A matrix for covariates}
}
\description{
EigenPrism 
Author: Lucas Janson (statweb.stanford.edu/~ljanson)
}
