% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorrelation.R
\name{empirical.cov.method}
\alias{empirical.cov.method}
\title{Historical data based method for covaraite decorrelation}
\usage{
empirical.cov.method(x, emp.sigma)
}
\arguments{
\item{x}{A matrix for covariates}

\item{emp.sigma}{A non-singular covariance matrix calculated from historical data}
}
\description{
Historical data based method for covaraite decorrelation
}
