% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2_methods.R
\name{GCTA.rr}
\alias{GCTA.rr}
\title{GCTA method}
\usage{
GCTA.rr(x, y, target = c("beta2", "h2")[1])
}
\arguments{
\item{x}{A matrix for covariates}

\item{y}{A continuous vector}

\item{target}{A character for estimation target}
}
\description{
GCTA method
}
