% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorrelation.R
\name{GLASSO.method}
\alias{GLASSO.method}
\title{GLASSO method for covaraite decorrelation}
\usage{
GLASSO.method(x, rho = 0.001)
}
\arguments{
\item{x}{A matrix for covariates}

\item{rho}{A penalty parameter}
}
\description{
GLASSO method for covaraite decorrelation
}
